Conflicts:
	src/generator_convolution_backward_avx2.c
	src/generator_convolution_forward_avx2.c
	src/generator_convolution_weight_update_avx2.c

+++ samples/jit_matcopy/jit_matcopy.c
@@ -49,7 +49,7 @@ int main(int argc, char* argv[])
 
   printf("This is a tester for JIT matcopy kernels!\n");
   desc.m = (1 < argc ? atoi(argv[1]) : 16);
-  desc.n = (2 < argc ? atoi(argv[2]) : desc.m);
+  desc.n = (2 < argc ? ((unsigned int)atoi(argv[2])) : desc.m);
   desc.ldi = LIBXSMM_MAX((unsigned int)(3 < argc ? atoi(argv[3]) : 0), desc.m);
   desc.ldo = LIBXSMM_MAX((unsigned int)(4 < argc ? atoi(argv[4]) : 0), desc.m);
   desc.unroll_level = (unsigned char)(5 < argc ? atoi(argv[5]) : 1);
